enhancement: set the logic when switching - When logging out from either Worker Mode or User Mode, the app defaults back to User Mode after logout. - If the user does not log out, the app retains the mode they were in. - When switching from User Mode to Worker Mode, the app navigates directly to the Worker Profile Screen. - On app relaunch, it redirects the user to the Home Screen of the mode they were in before restarting the app. - Ensures a consistent and seamless experience for mode switching and app relaunch behavior. - Also created a composable for the profile screen of the two modes to avoid duplication of code
