+++ app/src/main/java/com/arygm/quickfix/ui/uiMode/appContentUI/AppContentNavGraph.kt
@@ -1,10 +1,13 @@
 package com.arygm.quickfix.ui.uiMode.appContentUI
 
 import android.graphics.Bitmap
+import android.util.Log
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.navigation.compose.NavHost
@@ -23,6 +26,7 @@ import com.arygm.quickfix.ui.navigation.NavigationActions
 import com.arygm.quickfix.ui.uiMode.appContentUI.navigation.AppContentRoute
 import com.arygm.quickfix.ui.uiMode.workerMode.WorkerModeNavGraph
 import com.arygm.quickfix.ui.userModeUI.UserModeNavHost
+import com.arygm.quickfix.utils.loadAppMode
 
 @Composable
 fun AppContentNavGraph(
@@ -35,14 +39,28 @@ fun AppContentNavGraph(
     rootNavigationActions: NavigationActions,
     modeViewModel: ModeViewModel,
     userPreferencesViewModel: PreferencesViewModelUserProfile,
-    currentAppMode: AppMode,
     workerViewModel: ProfileViewModel,
     categoryViewModel: CategoryViewModel,
     locationViewModel: LocationViewModel
 ) {
   val appContentNavController = rememberNavController()
   val appContentNavigationActions = remember { NavigationActions(appContentNavController) }
+    var currentAppMode by remember { mutableStateOf(AppMode.USER) }
+    Log.d("userContent", "Current App Mode is empty: $currentAppMode")
+    LaunchedEffect(Unit) {
+        Log.d("userContent", "Loading App Mode")
+        currentAppMode =
+            when (loadAppMode(preferencesViewModel)) {
+                "USER" -> AppMode.USER
+                "WORKER" -> AppMode.WORKER
+                else -> {
+                    AppMode.WORKER
+                }
+            }
+    }
 
+    modeViewModel.switchMode(currentAppMode)
+    Log.d("MainActivity", "$currentAppMode")
   val startDestination =
       when (currentAppMode) {
         AppMode.USER -> AppContentRoute.USER_MODE
@@ -78,7 +96,14 @@ fun AppContentNavGraph(
 
         composable(AppContentRoute.WORKER_MODE) {
           WorkerModeNavGraph(
-              modeViewModel, isOffline, appContentNavigationActions, preferencesViewModel)
+              modeViewModel,
+              isOffline,
+              appContentNavigationActions,
+              preferencesViewModel,
+              accountViewModel,
+              rootNavigationActions,
+              userPreferencesViewModel
+          )
         }
       }
 }
