+++ app/src/main/java/com/arygm/quickfix/ui/uiMode/appContentUI/workerMode/WorkerModeNavGraph.kt
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -17,19 +18,23 @@ import androidx.compose.ui.platform.testTag
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.arygm.quickfix.model.account.AccountViewModel
 import com.arygm.quickfix.model.offline.small.PreferencesViewModel
+import com.arygm.quickfix.model.offline.small.PreferencesViewModelUserProfile
 import com.arygm.quickfix.model.switchModes.ModeViewModel
 import com.arygm.quickfix.ui.elements.QuickFixOfflineBar
 import com.arygm.quickfix.ui.navigation.BottomNavigationMenu
 import com.arygm.quickfix.ui.navigation.NavigationActions
+import com.arygm.quickfix.ui.profile.AccountConfigurationScreen
+import com.arygm.quickfix.ui.uiMode.appContentUI.userModeUI.profile.WorkerProfileScreen
 import com.arygm.quickfix.ui.uiMode.appContentUI.workerMode.announcements.AnnouncementsScreen
 import com.arygm.quickfix.ui.uiMode.appContentUI.workerMode.messages.MessagesScreen
-import com.arygm.quickfix.ui.uiMode.appContentUI.workerMode.profile.ProfileScreen
 import com.arygm.quickfix.ui.uiMode.workerMode.home.HomeScreen
 import com.arygm.quickfix.ui.uiMode.workerMode.navigation.WORKER_TOP_LEVEL_DESTINATIONS
 import com.arygm.quickfix.ui.uiMode.workerMode.navigation.WorkerRoute
 import com.arygm.quickfix.ui.uiMode.workerMode.navigation.WorkerScreen
 import com.arygm.quickfix.ui.uiMode.workerMode.navigation.getBottomBarIdWorker
+import com.arygm.quickfix.ui.userModeUI.navigation.UserScreen
 import kotlinx.coroutines.delay
 
 @Composable
@@ -37,13 +42,17 @@ fun WorkerModeNavGraph(
     modeViewModel: ModeViewModel,
     isOffline: Boolean,
     appContentNavigationActions: NavigationActions,
-    preferencesViewModel: PreferencesViewModel
+    preferencesViewModel: PreferencesViewModel,
+    accountViewModel: AccountViewModel,
+    rootMainNavigationActions: NavigationActions,
+    userPreferencesViewModel: PreferencesViewModelUserProfile
 ) {
   val workerNavController = rememberNavController()
   val workerNavigationActions = remember { NavigationActions(workerNavController) }
   var currentScreen by remember { mutableStateOf<String?>(null) }
   val shouldShowBottomBar by remember { derivedStateOf { true } }
-
+    val startDestination by modeViewModel.onSwitchStartDestWorker.collectAsState()
+    workerNavigationActions.setCurrentRoute(startDestination)
   var showBottomBar by remember { mutableStateOf(false) }
 
   // Delay the appearance of the bottom bar
@@ -76,7 +85,7 @@ fun WorkerModeNavGraph(
       }) { innerPadding ->
         NavHost(
             navController = workerNavigationActions.navController,
-            startDestination = WorkerRoute.HOME,
+            startDestination = startDestination,
             modifier = Modifier.padding(innerPadding)) {
               composable(WorkerRoute.HOME) { HomeNavHost(onScreenChange = { currentScreen = it }) }
               composable(WorkerRoute.MESSAGES) {
@@ -90,7 +99,7 @@ fun WorkerModeNavGraph(
                     onScreenChange = { currentScreen = it },
                     appContentNavigationActions,
                     preferencesViewModel,
-                    modeViewModel)
+                    modeViewModel, accountViewModel, workerNavigationActions, rootMainNavigationActions, userPreferencesViewModel)
               }
             }
       }
@@ -145,7 +154,11 @@ fun ProfileNavHost(
     onScreenChange: (String) -> Unit,
     appContentNavigationActions: NavigationActions,
     preferencesViewModel: PreferencesViewModel,
-    modeViewModel: ModeViewModel
+    modeViewModel: ModeViewModel,
+    accountViewModel: AccountViewModel,
+    workerNavigationActions: NavigationActions,
+    rootMainNavigationActions: NavigationActions,
+    userPreferencesViewModel: PreferencesViewModelUserProfile
 ) {
   val profileNavController = rememberNavController()
   val profileNavigationActions = remember { NavigationActions(profileNavController) }
@@ -155,11 +168,13 @@ fun ProfileNavHost(
   }
   NavHost(navController = profileNavController, startDestination = WorkerScreen.PROFILE) {
     composable(WorkerScreen.PROFILE) {
-      ProfileScreen(
-          preferencesViewModel,
-          modeViewModel,
-          appContentNavigationActions,
+      WorkerProfileScreen(
+          profileNavigationActions,
+          rootMainNavigationActions,preferencesViewModel,userPreferencesViewModel, appContentNavigationActions, modeViewModel
       )
     }
+      composable(WorkerScreen.ACCOUNT_CONFIGURATION) {
+          AccountConfigurationScreen(profileNavigationActions, accountViewModel, preferencesViewModel)
+      }
   }
 }
