+++ app/src/main/java/com/arygm/quickfix/ui/uiMode/appContentUI/workerMode/navigation/WorkerNavigation.kt
@@ -20,6 +20,7 @@ object WorkerScreen {
   const val ANNOUNCEMENT = "Announcement Screen"
   const val MESSAGES = "Messages Screen"
   const val PROFILE = "Profile Screen"
+    const val ACCOUNT_CONFIGURATION = "Account configuration Screen"
 }
 
 object WorkerTopLevelDestinations {
